Test if manufacturer overrides work
